refactor: merge ImagePostCard into PostCard
